[cartesian-match-tree] Use unsigned short instead of unsigned int for bitsets.

+++ src/search/per_state_bitset.h
@@ -8,7 +8,7 @@
 
 class BitsetMath {
 public:
-    using Block = unsigned int;
+    using Block = unsigned short;
     static_assert(
         !std::numeric_limits<Block>::is_signed,
         "Block type must be unsigned");
